Cambio en login, para trabajar con un solo panel. cambio en el boton cerrar

+++ El_Unico_Grupo3/El_Unico_Grupo3/FrmUsuarios.cs
@@ -17,26 +17,26 @@ namespace El_Unico_Grupo3
         private string Contraseña;
         private string TipoUsuario;
         private string Consulta;
-
+        public string CapturarTipoUsuario;
         ConexionDataBase conexionDB = new ConexionDataBase();
         public FrmUsuarios()
         {
             InitializeComponent();
-
+            lblCapturaTipoUsuario.Text = CapturarTipoUsuario;
         }
 
         private void btnRegistrar_Click(object sender, EventArgs e)
         {
 
-            if(EstaValidado())
+            if (EstaValidado())
             {
                 usuario = txtUsuario.Text;
                 Contraseña = txtContraseña.Text;
                 TipoUsuario = cbxTipoUsuario.SelectedItem.ToString();
-                Consulta = "INSERT INTO tab_usuario (Nombre_Usuario, Contrasena_Usuario, Tipo_Usuario) VALUES ('" +usuario +"', '"+ Contraseña + "', '" + TipoUsuario + "');";
+                Consulta = "INSERT INTO tab_usuario (Nombre_Usuario, Contrasena_Usuario, Tipo_Usuario) VALUES ('" + usuario + "', '" + Contraseña + "', '" + TipoUsuario + "');";
                 if (conexionDB.Insertar(Consulta))
                 {
-                    MessageBox.Show("Registro guardato con exito","Información", MessageBoxButtons.OK,MessageBoxIcon.Information);
+                    MessageBox.Show("Registro guardato con exito", "Información", MessageBoxButtons.OK, MessageBoxIcon.Information);
                     dgvUsuarios.DataSource = conexionDB.LlenarGrid("Select * from tab_usuario");
                 }
                 else
@@ -63,14 +63,14 @@ namespace El_Unico_Grupo3
                 NoError = false;
             }
 
-            if(txtContraseña.Text != txtConfirmarContraseña.Text)
+            if (txtContraseña.Text != txtConfirmarContraseña.Text)
             {
                 errorIConRegistroUser.SetError(txtConfirmarContraseña, "Contraseña no coincide");
                 NoError = false;
 
             }
 
-            if(cbxTipoUsuario.SelectedItem == null)
+            if (cbxTipoUsuario.SelectedItem == null)
             {
                 errorIConRegistroUser.SetError(cbxTipoUsuario, "Seleccione el tipo de usuario");
                 NoError = false;
@@ -90,7 +90,12 @@ namespace El_Unico_Grupo3
 
         private void FrmUsuarios_Load(object sender, EventArgs e)
         {
-            dgvUsuarios.DataSource = conexionDB.LlenarGrid("Select * from tab_usuario");
+            //comboBox1.DataSource = conexionDB.LlenarGrid("Select * from tab_producto");
+            //comboBox1.DisplayMember = "Nombre_Producto";
+            //comboBox1.ValueMember = "Id_Producto";
+            //MessageBox.Show(CapturarTipoUsuario);
+            //dgvUsuarios.DataSource = conexionDB.LlenarGrid("Select * from tab_usuario");
+            //lblCapturaTipoUsuario.Text = CapturarTipoUsuario;
         }
 
         private void txtIdUsuario_TextChanged(object sender, EventArgs e)
@@ -122,7 +127,7 @@ namespace El_Unico_Grupo3
         {
             if (EstaValidado())
             {
-                Consulta = "Update tab_usuario SET Nombre_Usuario='" + txtUsuario.Text + "', Contrasena_Usuario='" + txtContraseña.Text + "', Tipo_Usuario='"+cbxTipoUsuario.SelectedItem.ToString()+ "' where Id_Usuario=" + txtIdUsuario.Text;
+                Consulta = "Update tab_usuario SET Nombre_Usuario='" + txtUsuario.Text + "', Contrasena_Usuario='" + txtContraseña.Text + "', Tipo_Usuario='" + cbxTipoUsuario.SelectedItem.ToString() + "' where Id_Usuario=" + txtIdUsuario.Text;
                 if (conexionDB.Actualizar(Consulta))
                 {
                     MessageBox.Show("Registro actualizado con exito", "Información", MessageBoxButtons.OK, MessageBoxIcon.Information);
@@ -150,26 +155,33 @@ namespace El_Unico_Grupo3
 
         private void btnElimar_Click(object sender, EventArgs e)
         {
-            Consulta = "Delete from tab_usuario where Id_Usuario=" + txtIdUsuario.Text;
-            if (conexionDB.Eliminar(Consulta))
+            if (CapturarTipoUsuario == "Administrador")
             {
-                MessageBox.Show("Registro eliminado con exito", "Información", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                dgvUsuarios.DataSource = conexionDB.LlenarGrid("Select * from tab_usuario");
-                txtUsuario.Clear();
-                txtContraseña.Clear();
-                txtIdUsuario.Clear();
-                txtIdUsuarioBusqueda.Clear();
-                cbxTipoUsuario.SelectedItem = null;
-                txtUsuario.Focus();
+                Consulta = "Delete from tab_usuario where Id_Usuario=" + txtIdUsuario.Text;
+                if (conexionDB.Eliminar(Consulta))
+                {
+                    MessageBox.Show("Registro eliminado con exito", "Información", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    dgvUsuarios.DataSource = conexionDB.LlenarGrid("Select * from tab_usuario");
+                    txtUsuario.Clear();
+                    txtContraseña.Clear();
+                    txtIdUsuario.Clear();
+                    txtIdUsuarioBusqueda.Clear();
+                    cbxTipoUsuario.SelectedItem = null;
+                    txtUsuario.Focus();
+                }
+                else
+                {
+                    MessageBox.Show("No se pudo eliminar el registro", "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    txtUsuario.Clear();
+                    txtContraseña.Clear();
+                    txtIdUsuario.Clear();
+                    cbxTipoUsuario.SelectedItem = null;
+                    txtUsuario.Focus();
+                }
             }
             else
             {
-                MessageBox.Show("No se pudo eliminar el registro", "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                txtUsuario.Clear();
-                txtContraseña.Clear();
-                txtIdUsuario.Clear();
-                cbxTipoUsuario.SelectedItem = null;
-                txtUsuario.Focus();
+                MessageBox.Show("No tinene permiso de administrador");
             }
         }
 
@@ -177,6 +189,62 @@ namespace El_Unico_Grupo3
         {
             this.Close();
         }
-        
+
+        private void SelectIndex(object sender, EventArgs e)
+        {
+            //string ObtenerPrecio = "select PrecioVentaUnitario_Inventario from tab_inventario where FK_Producto_Inventario = " + (comboBox1.SelectedIndex = int.Parse(comboBox1.ValueMember));
+            //txtIdUsuarioBusqueda.Text = conexionDB.PrecioVenta(ObtenerPrecio);
+        }
+
+        private void btnAdministrar_Click(object sender, EventArgs e)
+        {
+            lblCapturaTipoUsuario.Text = CapturarTipoUsuario ;
+            if (lblCapturaTipoUsuario.Text == "Administrador") 
+            {
+
+                dgvUsuarios.Visible = true;
+                dgvUsuarios.DataSource = conexionDB.LlenarGrid("Select * from tab_usuario");
+                btnRegistrar.Visible = true;
+                btnElimar.Visible = true;
+                btnActualizar.Visible = true;
+                btnBuscarPorID.Visible = true;
+                btnCerrar.Visible = true;
+                txtIdUsuario.Visible = true;
+                txtUsuario.Visible = true;
+                txtContraseña.Visible = true;
+                txtConfirmarContraseña.Visible = true;
+                txtIdUsuarioBusqueda.Visible = true;
+                cbxTipoUsuario.Visible = true;
+                lblConfirmarContraseña.Visible = true;
+                lblContraseña.Visible = true;
+                lblIdBusquedaUsuario.Visible = true;
+                lblIdUsuario.Visible = true;
+                lblTipoUsuario.Visible = true;
+                lblUsuario.Visible = true;
+            }
+            else 
+            {
+                dgvUsuarios.Visible = true;
+                btnActualizar.Visible = true;
+                btnBuscarPorID.Visible = true;
+                btnCerrar.Visible = true;
+                txtIdUsuario.Visible = true;
+                txtUsuario.Visible = true;
+                txtContraseña.Visible = true;
+                txtConfirmarContraseña.Visible = true;
+                txtIdUsuarioBusqueda.Visible = true;
+                cbxTipoUsuario.Visible = true;
+                lblConfirmarContraseña.Visible = true;
+                lblContraseña.Visible = true;
+                lblIdBusquedaUsuario.Visible = true;
+                lblIdUsuario.Visible = true;
+                lblTipoUsuario.Visible = true;
+                lblUsuario.Visible = true;
+            }
     }
+       
+    }
+    
+        
+    
 }
